Moved PMOD connectors to be better centered vs metric mechanics.
Also referenced the 3d model that is part of pkl now instead of my local
model.

Added symbols for grounding mounting holes to the schematic. Did not add
them to layout yet as they need quite a lot of space on the board. Still
need to decide if we actually will do that. :/