Use normally closed jumpers. And fix VCCIO2 bypass. Closes #11
These jumpers connect two nets without impacting the schematic net
lists. This is a safer solution and easier to read than having to
remember where the net jumpers have to go and not miss a side path.
These jumpers use a hack of KiCAD where the DRC ignores lines on the
copper layer in footprints. The two pads of the jumper are shorted
within the footprint but they are not from the perspective of the DRC.
While I was already at it I have also replaced / separator with |
separator for multifunction nets. Unfortunately KiCAD reserves the /
separator for its multisheet net naming. This patch should fix the
bypass of the VCCIO0 jumper that was a mistake in the previous versions
of the board.