Add an FPGA CRESET jumper. Fixes #9
Adding an in line resistor for the CRESET line from FTDI as well as a
jumper that allows access and pulldown of the CRESET line. This allows
us to override the CRESET signal from FTDI and force the FPGA chip to
stay in reset in cases when we need to access the flash and prevent the
FPGA from configuring the FLASH chip. Especially in cases when the FPGA
is using the FLASH after the design is loaded and configures the flash
into QSPI and other higher bit modes.