Initial commit of iCEBreaker V1.0b.
* Added 1V8 jumper for FPGA GPIO bank 2. Not added to 0&1 as the pins on
that bank are also connected to the FTDI chip. Extra precausions have to
be taken in that situation as the FTDI chip GPIO are connected to 3.3V.
The user will have to assure that there are no problems and the FPGA
will not fry either itself or the FTDI chip. We do not want to encourage
that, but if you feel adventerous you are welcome to try. ;) Closes #15