Added easier way for adding "piggyback" SPI RAM.
Added a pad near the SPI flash CS pin. This pad is connected to ~LEDG.
This will allow us to solder on an SPI RAM chip on top of the SPI flash
chip and connect it's CS line to ~LEDG signal. Allowing us to use the
flash and RAM through the same SPI interface.

This closes #19 for the V1.0c milestone. We will revisit this topic for
V2.0 milestone.