Added "ear" circuits. WS2812 & Shitty Addon.
The "top" ear is a RGB LED connector to WS2812 adapter with level
shifting to 5V. It supports one WS2812 string interface. (there is not
enough space for more channels unfortunately.

The "bottom" ear is a RGB LED connector to Shitty Addon adapter with I2C
pullups. This allows for connecting badgelife shitty addon mini badges
to the iCEBreaker.